Extract GroupByHash interface
