+++ presto-main/src/test/java/com/facebook/presto/operator/TestGroupByHash.java
@@ -25,6 +25,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
+import static com.facebook.presto.operator.GroupByHash.createGroupByHash;
 import static com.facebook.presto.spi.type.BigintType.BIGINT;
 import static com.facebook.presto.spi.type.DoubleType.DOUBLE;
 import static com.facebook.presto.spi.type.VarcharType.VARCHAR;
@@ -40,7 +41,7 @@ public class TestGroupByHash
     public void testAddPage()
             throws Exception
     {
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(BIGINT), new int[] { 0 }, Optional.of(1), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(BIGINT), new int[] {0}, Optional.of(1), 100);
         for (int tries = 0; tries < 2; tries++) {
             for (int value = 0; value < MAX_GROUP_ID; value++) {
                 Block block = BlockAssertions.createLongsBlock(value);
@@ -68,7 +69,7 @@ public class TestGroupByHash
     public void testGetGroupIds()
             throws Exception
     {
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(BIGINT), new int[] { 0 }, Optional.of(1), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(BIGINT), new int[] {0}, Optional.of(1), 100);
         for (int tries = 0; tries < 2; tries++) {
             for (int value = 0; value < MAX_GROUP_ID; value++) {
                 Block block = BlockAssertions.createLongsBlock(value);
@@ -89,7 +90,7 @@ public class TestGroupByHash
     public void testTypes()
             throws Exception
     {
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(VARCHAR), new int[] { 0 }, Optional.of(1), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(VARCHAR), new int[] {0}, Optional.of(1), 100);
         // Additional bigint channel for hash
         assertEquals(groupByHash.getTypes(), ImmutableList.of(VARCHAR, BIGINT));
     }
@@ -100,7 +101,7 @@ public class TestGroupByHash
     {
         Block valuesBlock = BlockAssertions.createStringSequenceBlock(0, 100);
         Block hashBlock = TypeUtils.getHashBlock(ImmutableList.of(VARCHAR), valuesBlock);
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(VARCHAR), new int[] { 0 }, Optional.of(1), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(VARCHAR), new int[] {0}, Optional.of(1), 100);
 
         GroupByIdBlock groupIds = groupByHash.getGroupIds(new Page(valuesBlock, hashBlock));
         for (int i = 0; i < groupIds.getPositionCount(); i++) {
@@ -134,7 +135,7 @@ public class TestGroupByHash
         Block valuesBlock = BlockAssertions.createLongsBlock(values);
         Block hashBlock = TypeUtils.getHashBlock(ImmutableList.of(BIGINT), valuesBlock);
 
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(BIGINT), new int[] { 0 }, Optional.of(1), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(BIGINT), new int[] {0}, Optional.of(1), 100);
         groupByHash.getGroupIds(new Page(valuesBlock, hashBlock));
         assertEquals(groupByHash.getGroupCount(), 50);
 
@@ -154,7 +155,7 @@ public class TestGroupByHash
     {
         Block valuesBlock = BlockAssertions.createDoubleSequenceBlock(0, 10);
         Block hashBlock = TypeUtils.getHashBlock(ImmutableList.of(DOUBLE), valuesBlock);
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(DOUBLE), new int[] { 0 }, Optional.of(1), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(DOUBLE), new int[] {0}, Optional.of(1), 100);
         groupByHash.getGroupIds(new Page(valuesBlock, hashBlock));
 
         Block testBlock = BlockAssertions.createDoublesBlock((double) 3);
@@ -173,7 +174,7 @@ public class TestGroupByHash
         Block valuesBlock = BlockAssertions.createDoubleSequenceBlock(0, 10);
         Block stringValuesBlock = BlockAssertions.createStringSequenceBlock(0, 10);
         Block hashBlock = TypeUtils.getHashBlock(ImmutableList.of(DOUBLE, VARCHAR), valuesBlock, stringValuesBlock);
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(DOUBLE, VARCHAR), new int[] { 0, 1 }, Optional.of(2), 100);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(DOUBLE, VARCHAR), new int[] {0, 1}, Optional.of(2), 100);
         groupByHash.getGroupIds(new Page(valuesBlock, stringValuesBlock, hashBlock));
 
         Block testValuesBlock = BlockAssertions.createDoublesBlock((double) 3);
@@ -191,7 +192,7 @@ public class TestGroupByHash
         Block hashBlock = TypeUtils.getHashBlock(ImmutableList.of(VARCHAR), valuesBlock);
 
         // Create group by hash with extremely small size
-        GroupByHash groupByHash = new GroupByHash(ImmutableList.of(VARCHAR), new int[] { 0 }, Optional.of(1), 4);
+        GroupByHash groupByHash = createGroupByHash(ImmutableList.of(VARCHAR), new int[] {0}, Optional.of(1), 4);
         groupByHash.getGroupIds(new Page(valuesBlock, hashBlock));
 
         // Ensure that all groups are present in group by hash
